iOS bug fix.  Android restore receiver enhancement
- Fix iOS schedule method to remove existing notification first.
- Refactor Android receivers.
- Create a base receiver with the logic for triggering a notification.
- Extend that receiver for both restore and trigger.
- Use same notification logic on boot as you do during a normal trigger.
