+++ src/android/RestoreReceiver.java
@@ -34,13 +34,15 @@ import de.appplant.cordova.plugin.notification.Notification;
 import de.appplant.cordova.plugin.notification.Request;
 import de.appplant.cordova.plugin.notification.receiver.AbstractRestoreReceiver;
 
+import static de.appplant.cordova.plugin.localnotification.LocalNotification.fireEvent;
+import static de.appplant.cordova.plugin.localnotification.LocalNotification.isAppRunning;
+
 /**
  * This class is triggered upon reboot of the device. It needs to re-register
  * the alarms with the AlarmManager since these alarms are lost in case of
  * reboot.
  */
 public class RestoreReceiver extends AbstractRestoreReceiver {
-
     /**
      * Called when a local notification need to be restored.
      *
@@ -53,7 +55,7 @@ public class RestoreReceiver extends AbstractRestoreReceiver {
         boolean after = date != null && date.after(new Date());
 
         if (!after && toast.isHighPrio()) {
-            toast.show();
+            performNotification(toast);
         } else {
             toast.clear();
         }
@@ -64,6 +66,16 @@ public class RestoreReceiver extends AbstractRestoreReceiver {
         if (after || toast.isRepeating()) {
             mgr.schedule(request, TriggerReceiver.class);
         }
+    }
+
+    @Override
+    public void dispatchAppEvent(String key, Notification notification) {
+        fireEvent(key, notification);
+    }
+
+    @Override
+    public boolean checkAppRunning() {
+        return isAppRunning();
     }
 
     /**
